passed two first tests in wave 6
