+++ app/__init__.py
@@ -33,7 +33,7 @@ def create_app(test_config=None):
     from .routes import tasks_bp
     app.register_blueprint(tasks_bp)
 
-    from .routes import goals_bp
+    from .goal_routes import goals_bp
     app.register_blueprint(goals_bp)
 
     return app
