+++ app/models/goal.py
@@ -1,7 +1,6 @@
 from app import db
 
-
 class Goal(db.Model):
     goal_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     title = db.Column(db.String)
-
+    tasks = db.relationship("Task", back_populates="goal", lazy=True)
